pyproject.toml
[project]
name = "pyobsidianrpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "jupyter>=1.1.1",
  "pandas>=2.3.3",
  "pandoc>=2.4",
  "pyyaml>=6.0.3",
]


[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
